+++ src/main.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   main.c                                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: pruenrua <pruenrua@student.42.fr>          +#+  +:+       +#+        */
+/*   By: wave <wave@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/10 21:25:08 by pruenrua          #+#    #+#             */
-/*   Updated: 2024/06/07 20:04:38 by pruenrua         ###   ########.fr       */
+/*   Updated: 2024/06/09 12:58:03 by wave             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,6 +27,7 @@ int	main(int ac, char **av)
 	d->parser_data = main_parser(av[1]);
 	if (d->parser_data == NULL)
 		return (puterror("Invalid Data in file"), terminate_and_cleanup(&d), 1);
+	//print_map_data(d->parser_data);
 	if (!set_from_parser_data(d))
 		return (terminate_and_cleanup(&d), 1);
 	// if (prepare_render(d) == false)
