+++ src/parser/get_texture_file.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   get_texture_file.c                                 :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: pruenrua <pruenrua@student.42.fr>          +#+  +:+       +#+        */
+/*   By: wave <wave@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/21 18:40:55 by pruenrua          #+#    #+#             */
-/*   Updated: 2024/06/07 23:07:22 by pruenrua         ###   ########.fr       */
+/*   Updated: 2024/06/09 12:54:15 by wave             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -52,7 +52,7 @@ int	checklist(t_parser_data *res, char **data)
 char	*read_loop(int read_co, char *chdata, int fd, char *buffer)
 {
 	char	*tmp;
-
+	dprintf(2, "start read loop!!!\n");
 	while (read_co > 0)
 	{
 		tmp = chdata;
@@ -60,19 +60,22 @@ char	*read_loop(int read_co, char *chdata, int fd, char *buffer)
 		buffer[read_co] = '\0';
 		chdata = ft_strjoin(tmp, buffer);
 		if (!chdata)
-			return (puterror(": Fatal"), free(buffer), NULL);
-		// new check
-		if (false == ft_isprint_mk2(buffer[0]))
-			read_co = 0;
-		// tam maiiiii
+			return (puterror(": Fatal"), free(buffer), free(tmp), NULL);
+		// new check will read until the non printable char are occur and '\0'
+		// deprecate from
+		// vvvvvvvv
 		// if ((buffer[0] < 32 || buffer[0] > 126) \
 		// 	&& buffer[0] != '\n' && read_co > 0)
 		// 	read_co = 0;
+		if (false == ft_isprint_mk2(buffer[0]))
+			read_co = 0;
 		free(tmp);
 	}
 	free(buffer);
+	dprintf(2, "read is \n################\n%s\n##############\nreturn\n", chdata);
 	if (read_co < -1)
 		perror("FILE READER : ");
+	
 	return (chdata);
 }
