+++ src/parser/parser.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   parser.c                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: pruenrua <pruenrua@student.42.fr>          +#+  +:+       +#+        */
+/*   By: wave <wave@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/13 15:10:32 by tpoungla          #+#    #+#             */
-/*   Updated: 2024/06/07 23:04:47 by pruenrua         ###   ########.fr       */
+/*   Updated: 2024/06/09 12:59:32 by wave             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -37,6 +37,7 @@ char	**file_reader(int fd)
 	char	*buffer;
 	int		read_count;
 
+	dprintf(2, "start file reader !!!\n");
 	read_count = 1;
 	buffer = ft_calloc(sizeof(char), (BUFFER_SIZE + 1));
 	if (!buffer)
@@ -46,9 +47,6 @@ char	**file_reader(int fd)
 	if (!chdata)
 		return (free(buffer), puterror(": Fatal"), NULL);
 	chdata = read_loop(read_count, chdata, fd, buffer);
-	// for (int ii = 0; chdata[ii]; ii++)
-	dprintf(2, "----------------------------------------------------------\n");
-	dprintf(2, "-> [%s]\n", chdata);
 	// if (scanner(chdata) == 0 || !ft_strlen(chdata))
 	// 	return (free(chdata), NULL);
 	map = ft_split(chdata, '\n');
@@ -56,34 +54,6 @@ char	**file_reader(int fd)
 	return (map);
 }
 
-t_parser_data	*main_parser(char *file_name)
-{
-	t_parser_data	*res;
-	int				fd;
-	char			**data;
-
-	res = ft_calloc(sizeof(t_parser_data), 1);
-	if (!is_file_valid(file_name, ".cub"))
-		return (free(res), NULL);
-	fd = open(file_name, O_RDONLY);
-	if (fd < 0)
-		return (free(res), NULL);
-	data = file_reader(fd);
-	close(fd);
-	dprintf(2, "file reader done!!\n");
-	if (!checklist(res, data))
-		return (NULL);
-	res->height = find_height(data);
-	res->width = find_width(data);
-	if (res->height == 0 || res->width == 0)
-		return (free_2dwithres(res, data), NULL);
-	res->maps_data = init_map(data, find_width(data), find_height(data));
-	if (!scan4player(res->maps_data))
-		return (free_parser(res), NULL);
-	if (!border_checker(res, res->maps_data))
-		return (free_parser(res), NULL);
-	return (res);
-}
 
 void	free_2dwithres(t_parser_data *res, char **data)
 {
